Struct tokio_uds::UnixListener

pub struct UnixListener { /* fields omitted */ }

A Unix socket which can accept connections from other Unix sockets.

Implementations

impl UnixListener

pub fn bind<P>(path: P) -> Result<UnixListener> where
P: AsRef<Path>,

Creates a new UnixListener bound to the specified path.

pub fn from_std(listener: UnixListener, handle: &Handle) -> Result<UnixListener>

Consumes a UnixListener in the standard library and returns a nonblocking UnixListener from this crate.

The returned listener will be associated with the given event loop specified by handle and is ready to perform I/O.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local socket address of this listener.

pub fn poll_read_ready(&self, ready: Ready) -> Poll<Ready, Error>

Test whether this socket is ready to be read or not.

pub fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

pub fn poll_accept(&self) -> Poll<(UnixStream, SocketAddr), Error>

Attempt to accept a connection and create a new connected UnixStream if successful.

This function will attempt an accept operation, but will not block waiting for it to complete. If the operation would block then a “would block” error is returned. Additionally, if this method would block, it registers the current task to receive a notification when it would otherwise not block.

Note that typically for simple usage it’s easier to treat incoming connections as a Stream of UnixStreams with the incoming method below.

Panics

This function will panic if it is called outside the context of a future’s task. It’s recommended to only call this from the implementation of a Future::poll, if necessary.

pub fn poll_accept_std(&self) -> Poll<(UnixStream, SocketAddr), Error>

Attempt to accept a connection and create a new connected UnixStream if successful.

This function is the same as poll_accept above except that it returns a mio_uds::UnixStream instead of a tokio_udp::UnixStream. This in turn can then allow for the stream to be associated with a different reactor than the one this UnixListener is associated with.

This function will attempt an accept operation, but will not block waiting for it to complete. If the operation would block then a “would block” error is returned. Additionally, if this method would block, it registers the current task to receive a notification when it would otherwise not block.

Note that typically for simple usage it’s easier to treat incoming connections as a Stream of UnixStreams with the incoming method below.

Panics

This function will panic if it is called outside the context of a future’s task. It’s recommended to only call this from the implementation of a Future::poll, if necessary.

pub fn incoming(self) -> Incoming

Consumes this listener, returning a stream of the sockets this listener accepts.

This method returns an implementation of the Stream trait which resolves to the sockets the are accepted on this listener.