Crate tokio_uds

Unix Domain Sockets for Tokio.

Note: This crate is deprecated in tokio 0.2.x and has been moved into tokio::uds behind the uds feature flag.

This crate provides APIs for using Unix Domain Sockets with Tokio.

Structs

ConnectFuture
Future returned by UnixStream::connect which will resolve to a UnixStream when the stream is connected.

Incoming
Stream of listeners

RecvDgram
A future for receiving datagrams from a Unix datagram socket.

SendDgram
A future for writing a buffer to a Unix datagram socket.

UCred
Credentials of a process

UnixDatagram
An I/O object representing a Unix datagram socket.

UnixDatagramFramed
A unified Stream and Sink interface to an underlying UnixDatagram, using the Encoder and Decoder traits to encode and decode frames.

UnixListener
A Unix socket which can accept connections from other Unix sockets.

UnixStream
A structure representing a connected Unix socket.