Crate tokio_uds

Bindings for Unix Domain Sockets and futures

This crate provides bindings between mio_uds, the mio crate for Unix Domain sockets, and futures. The APIs and bindings in this crate are very similar to the TCP and UDP bindings in the futures-mio crate. This crate is also an empty crate on Windows, as Unix Domain Sockets are Unix-specific.

Structs

RecvDgram
A future for receiving datagrams from a Unix datagram socket.

SendDgram
A future for writing a buffer to a Unix datagram socket.

UCred
Credentials of a process

UnixDatagram
An I/O object representing a Unix datagram socket.

UnixDatagramFramed
A unified Stream and Sink interface to an underlying UnixDatagramSocket, using the UnixDatagramCodec trait to encode and decode frames.

UnixListener
A Unix socket which can accept connections from other unix sockets.

UnixStream
A structure representing a connected unix socket.

Traits

UnixDatagramCodec
Encoding of frames via buffers.