Struct tokio_uds::UnixListener

pub struct UnixListener { /* fields omitted */ }

A Unix socket which can accept connections from other unix sockets.

Methods

impl UnixListener

pub fn bind<P>(path: P, handle: &Handle) -> Result<UnixListener> where
P: AsRef<Path>,

Creates a new UnixListener bound to the specified path.

pub fn from_listener(
listener: UnixListener,
handle: &Handle
) -> Result<UnixListener>

Consumes a UnixListener in the standard library and returns a nonblocking UnixListener from this crate.

The returned listener will be associated with the given event loop specified by handle and is ready to perform I/O.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local socket address of this listener.

pub fn poll_read(&self) -> Async<()>

Test whether this socket is ready to be read or not.

pub fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

pub fn accept(&mut self) -> Result<(UnixStream, SocketAddr)>

Attempt to accept a connection and create a new connected UnixStream if successful.

This function will attempt an accept operation, but will not block waiting for it to complete. If the operation would block then a "would block" error is returned. Additionally, if this method would block, it registers the current task to receive a notification when it would otherwise not block.

Note that typically for simple usage it's easier to treat incoming connections as a Stream of UnixStreams with the incoming method below.

Panics

This function will panic if it is called outside the context of a future's task. It's recommended to only call this from the implementation of a Future::poll, if necessary.

pub fn incoming(self) -> IoStream<(UnixStream, SocketAddr)>

Consumes this listener, returning a stream of the sockets this listener accepts.

This method returns an implementation of the Stream trait which resolves to the sockets the are accepted on this listener.