Struct tokio_uds::UnixStream

pub struct UnixStream { /* fields omitted */ }

A structure representing a connected unix socket.

This socket can be connected directly with UnixStream::connect or accepted from a listener with UnixListener::incoming. Additionally, a pair of anonymous Unix sockets can be created with UnixStream::pair.

Methods

impl UnixStream

pub fn connect<P>(p: P, handle: &Handle) -> Result<UnixStream> where
P: AsRef<Path>,

Connects to the socket named by path.

This function will create a new unix socket and connect to the path specified, performing associating the returned stream with the provided event loop's handle.

pub fn from_stream(stream: UnixStream, handle: &Handle) -> Result<UnixStream>

Consumes a UnixStream in the standard library and returns a nonblocking UnixStream from this crate.

The returned stream will be associated with the given event loop specified by handle and is ready to perform I/O.

pub fn pair(handle: &Handle) -> Result<(UnixStream, UnixStream)>

Creates an unnamed pair of connected sockets.

This function will create a pair of interconnected unix sockets for communicating back and forth between one another. Each socket will be associated with the event loop whose handle is also provided.

pub fn need_read(&self)

Indicates to this source of events that the corresponding I/O object is no longer readable, but it needs to be.

Panics

This function will panic if called outside the context of a future's task.

pub fn need_write(&self)

Indicates to this source of events that the corresponding I/O object is no longer writable, but it needs to be.

Panics

This function will panic if called outside the context of a future's task.

pub fn poll_read(&self) -> Async<()>

Test whether this socket is ready to be read or not.

pub fn poll_write(&self) -> Async<()>

Test whether this socket is ready to be written to or not.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address of the local half of this connection.

pub fn peer_addr(&self) -> Result<SocketAddr>

Returns the socket address of the remote half of this connection.

pub fn peer_cred(&self) -> Result<UCred>

Returns effective credentials of the process which called connect or socketpair.

pub fn take_error(&self) -> Result<Option<Error>>

Returns the value of the SO_ERROR option.

pub fn shutdown(&self, how: Shutdown) -> Result<()>

Shuts down the read, write, or both halves of this connection.

This function will cause all pending and future I/O calls on the specified portions to immediately return with an appropriate value (see the documentation of Shutdown).